+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+### Master
+
+Features:
+- Introduce `MissingUniqueIndexChecker` to find missing unique indexes.
+
 ### [0.4.0] - 2019/01/10
 
 Features:
-- Introduce `BelongsToPresenceChecker` to find missed foreign keys.
+- Introduce `BelongsToPresenceChecker` to find missing foreign keys.
 
 Breaking changes:
 - Rename `PresenceValidationChecker` to `ColumnPresenceChecker` for simplicity. Please update your configuration files properly.